+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * New API for answering a call inside an application:
 
   * PUT `/1.0/applications/{uuid}/calls/{call_id}/answer`
+  * PUT `/1.0/applications/{uuid}/calls/{call_id}/ring`
 
 
 ## 19.09